+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### February
 
+#### 22 February 2024
+
+* Load client configuration from file instead of hardcoding the values.
+
 #### 21 February 2024
 
 * Successfully integrated Dear ImGui with the renderer - now the client can have GUIs. Still need